Write a function that executes a function given as a parameter on each element of an array.

+++ 0x0E-function_pointers/1-array_iterator.c
@@ -1,13 +1,16 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator -
+ * array_iterator - Executes a function given as a parameter on each elem
+ * @array: array to print
+ * @size: size of array
+ * @action: the called function that print
  *
- * Return: 
  */
-void array_iterator(int *array, int size, void (*action)(int))
+void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	int x;
+
 	if (action != NULL)
 	{
 		for (x = 0; x < size; x++)
